requirements.txt
igdb-api-v4
requests
